tests/wpt/tests/css/mediaqueries/navigation-controls.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/fallaciousreasoning/backbutton-mediaquery/blob/master/explainer.md" />
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>
<script type="text/javascript" src="resources/matchmedia-utils.js"></script>
<script>
    query_should_be_known("(navigation-controls)");
    query_should_be_known("(navigation-controls: none)");
    query_should_be_known("(navigation-controls: back-button)");
    query_should_be_unknown("(navigation-controls: none back-button)");
    query_should_be_unknown("(navigation-controls: back-button/none)");
    test(() => {
        let booleanContext = window.matchMedia("(navigation-controls)");
        let none = window.matchMedia("(navigation-controls: none)");
        assert_equals(booleanContext.matches, !none.matches);
    }, "Check that none evaluates to false in the boolean context");
</script>
